backend/requirements.txt
fastapi
uvicorn[standard]  # Uvicorn 서버 및 표준 의존성 (gunicorn, httptools 등 포함 가능)
sqlalchemy
mysql-connector-python  # MySQL DB 연결 드라이버
pydantic
python-dotenv  # .env 파일 로드용
bcrypt  # 비밀번호 해싱용